+++ Runtime/Counter.cs
@@ -1,7 +1,10 @@
 using System;
+using System.Collections.Generic;
+using System.Collections.Specialized;
 using System.Runtime.CompilerServices;
 using System.Threading;
-
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 using UnityEngine;
 
 [assembly: InternalsVisibleTo("Unity.Simulation.Games.Tests")]
@@ -10,17 +13,24 @@ namespace Unity.Simulation.Games
     [Serializable]
     internal class Counter
     {
-        [SerializeField]
+        [JsonProperty]
         string _name;
 
+        [JsonIgnore]
         public string Name { get { return _name; } }
 
-        [SerializeField]
-        Int64 _value;
+        [JsonProperty]
+        internal Int64 _value;
         
-        [NonSerialized]
+        [JsonProperty]
+        internal OrderedDictionary _snapshots;
+
+        [JsonProperty]
+        internal StepSeries _stepSeries;
+
         internal Int64 _count;
 
+        [JsonIgnore]
         public Int64 Value { get { return _value; } }
 
         public Counter(string name)
@@ -37,6 +47,31 @@ namespace Unity.Simulation.Games
         internal void Reset(Int64 value = 0)
         {
             Interlocked.Exchange(ref _value, value);
+        }
+
+        internal void Snapshot(String label)
+        {
+            if (_snapshots == null)
+            {
+                _snapshots = new OrderedDictionary();
+            }
+            _snapshots.Add(label, _value);
+        }
+
+        internal void CaptureStepSeries(int intervalSeconds)
+        {
+            if (intervalSeconds <= 0)
+            {
+                Debug.LogError("Interval seconds must be greater than 0");
+                return;
+            }
+
+            if (_stepSeries != null)
+            {
+                Debug.LogError("Step Series has already been enabled for Counter " + _name);
+            }
+
+            _stepSeries = new StepSeries(StepSeriesInterval.Seconds, intervalSeconds, this);
         }
     }
 }