+++ Runtime/RemoteConfigProvider.cs
@@ -42,7 +42,7 @@ namespace Unity.Simulation.Games
                 string execution_id = Configuration.Instance.SimulationConfig.execution_id.Split(':')[2];
                 string definition_id = Configuration.Instance.SimulationConfig.definition_id.Split(':')[2];
 
-                Debug.Log("Instance id = " + Configuration.Instance.GetInstanceId().ToString()
+                Log.I("Instance id = " + Configuration.Instance.GetInstanceId().ToString()
                                            + " execution id - " + execution_id.ToString()
                                            + " defitnion id - " + definition_id.ToString());
                 var appParams = Configuration.Instance.GetAppParams<GameSimAppParams>();
@@ -70,12 +70,14 @@ namespace Unity.Simulation.Games
             switch (response.requestOrigin)
             {
                 case ConfigOrigin.Default:
-                    Debug.Log("No settings loaded this session; using default values.");
+                    Log.I("No settings loaded this session; using default values.");
                     break;
                 case ConfigOrigin.Cached:
-                    Debug.Log("No settings loaded this session; using cached values from a previous session.");
+                    Log.I("No settings loaded this session; using cached values from a previous session.");
                     break;
                 case ConfigOrigin.Remote:
+                    Log.I("Remote Config fetch was completed successfully with the server");
+                    Log.I("Config fetched: " + ConfigManager.appConfig.config.ToString(Newtonsoft.Json.Formatting.None));
                     GameSimManager.Instance.AddMetaData = () => ConfigManager.appConfig.config.ToString(Newtonsoft.Json.Formatting.None);
                     ConfigHandler?.Invoke(new GameSimConfigResponse());
                     break;