+++ Runtime/StepSeriesInterval.cs
@@ -1,0 +1,7 @@
+namespace Unity.Simulation.Games
+{
+    public enum StepSeriesInterval
+    {
+        Seconds
+    }
+}