+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.3-preview.1] - 2020-06-25
+
+## This is the v0.4.3-preview.1 release of *Unity Game Simulation*
+- Added support for step series metrics
+- Updated build name validation to match Unity Simulation
+
+## [0.4.2-preview.1] - 2020-05-05
+
+## This is the v0.4.2-preview.1 release of *Unity Game Simulation*
+- On build & upload will now set `PlayerSettings.runInBackground` to true, then set it back to what it was previously after the build is complete.
+- On build & upload in Unity versions <2019.1 will now set `PlayerSettings.displayResolutionDialog` to false, then set it back to what it was previously after the build is complete.
+- Added logging of values fetched from Remote Config to the player log for debugging purposes.
+- Updated Unity Simulation Core package to `0.0.10-preview.14` due to a fps bug in `0.0.10-preview.11`.
+
 ## [0.4.1-preview.7] - 2020-05-04
 
 ## This is the v0.4.1-preview.6 release of *Unity Game Simulation*